corección de erres nueva version

+++ src/main/java/ec/edu/ups/icc/fundamentos01/users/dtos/UserResponseDto.java
@@ -1,7 +1,10 @@
 package ec.edu.ups.icc.fundamentos01.users.dtos;
 
+/**
+ * DTO de respuesta para usuarios
+ */
 public class UserResponseDto {
-    public int id;
+    public Long id;
     public String name;
     public String email;
 }
